moved App.js to src/components
